+++ src/render/mod.rs
@@ -41,4 +41,4 @@ pub use options::{
 pub use style_mapping::StyleMapping;
 pub use text::to_text;
 #[cfg(feature = "refine")]
-pub use unrefine::{refine, RefineOptions};
+pub use unparser_shared::refine::{refine, RefineOptions};